+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 ## Next
 
 - Add functionality to remove `placeholder` <option> in BasicDropdown component
-- Upgrade Storybook to version 8 and update Svelte dependencies
+- Upgrade Storybook to version 8 
+- Upgrade SvelteKit to version 2 and update other Svelte-related dependencies
+- Upgrade to Vite version 5
 
 ## v0.6.0