+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 - Add functionality to remove `placeholder` <option> in BasicDropdown component
 - Turn off interaction events `onDestroy` of DatawrapperIframe
 - Add Toggle component
+- Upgrade Storybook to version 8 
+- Upgrade SvelteKit to version 2 and update other Svelte-related dependencies
+- Upgrade to Vite version 5
 
 ## v0.6.0