+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Upgrade SvelteKit to version 2 and update other Svelte-related dependencies
 - Upgrade to Vite version 5
 - Add Scrolly component
+- Update story ordering config
 
 ## v0.6.0